Make system-shutdown testable with unit tests
